extracted cases; tests pass
